:sparkles: [#149] Initial work on index document task
Ensure that the task is idempotent by only indexing when we can or
should index, and otherwise invoking the gpp-search client.

This task will be hooked up to the API endpoints and API change
events to ensure that mutations from either end result in a sync
operation to the index.