:sparkles: [#10] Ensure that documents are removed from index upon deleteion
When a document is deleted via the admin interface, it must
also be deleted from the search index. We can't pass the primary
key of the object to the celery task, because the DB lookup will
fail as it has been deleted prior to the celery task being picked
up.

Instead, we create a dummy instance and forcibly remove the object
via its UUID reference.