:sparkles: [GPP-Woo/GPP-Zoeken#12] Pass request object to inline document formset
We can pass the request object down to the inline formset so that
the absolute URL can be calculated and passed along to the index
celery task.

This addresses https://github.com/GPP-Woo/GPP-publicatiebank/pull/216/files\#r1984938432
and thus ensures we don't lose the full text indexed data of the
document content.