:package: [#232] Vendor MultipartEncoder from requests-toolbelt
requests-toolbelt has some useful utilities, but I opted to vendor te
code that's relevant to use instead of adding the dependency:

* it hasn't seen a release since two years
* there's a lot more stuff in the library that increases the attack
  surface, yet we don't make use of it
* github activity suggests it's barely maintained and users are asking
  for a new release, which has not been addressed since December last
  year

Minor modifications have been made to make the type checking pass,
as this codebase is entirely untyped.

Lastly, the core mechanic why this works is by implementing the
MultipartEncoder.read method, which by default (if no size is
specified) still loads the entire file/body into memory. This is
what you'll actually see happening when using VCR to record the
tests, as it needs to record the request bodies to replay/match
them anyway. These tests can therefore not provide any guarantees
about actually performing chunked uploads to Open Zaak :(

I disabled VCR for the test, spun up the Docker container for
Open Zaak and put a breakpoint on the read method to figure
out if and what causes small chunks to be read. I managed to
trace this back to urllib3 HTTPConnection which by default
reads body chunks of 16MB, which is definitely okay for us.