:zap: [#232] Leverage MultipartEncoder to send chunked/streamed file part uploads
Before this patch, the entire file had to be read in memory to upload
it to the Documenten API because that's what vanilla requests offers.
Streaming files is possible, but only if there's no multipart form
data involved too, and we need to send along the lock as well.

This is a problem in terms of container memory limits that eventually
get OOM-killed, and even increasing memory requests/limits provides
no guarantees as (multiple) users may upload parts concurrently, which
all add to the container memory consumption. And frankly, 4GB of mem
usage for a web app is prepostorous.

We had to look beyond vanilla requests - requests-toolkit ships with a
built-in class for multipart chunked/streamed uploads, so it was
vendored (see previous commit) and the client implementation now makes
use of this solution/approach.

Writing tests for this is tough, since it's pretty low-level HTTP and
from the sending or receiving end, nothing really changes. It has been
verified manually that it appears to work, and we'll test it for real
after deploying it in a memory-limited container.

Note that a number of mechanisms play together here:

* for not-small file uploads, Django will put the file upload on disk
  in a temporary file (rather than keep it in memory). The cut-off for
  this can be controlled via Django settings
* we pass around the upload file handle to the MultipartEncoder, which
  can grab the file size from the file handle and can read chunks of
  the file
* the MultipartEncoder ensures the boundaries are handled properly and
  finished the request
* the remote Documenten API receives the TCP/HTTP stream and there a
  similar file upload-handling process runs (large files are put on
  disk as a temp file etc.)

I'm not sure if this is doing anything special at the HTTP level, like
chunked encoding or anything like that, and I hope I will not need to
know.