:card_file_box: [#260] Fix data migration for old (dev) environments
Publications/documents may exist in the DB from before their creation
was tracked in the audit logs, as the audit logs feature was built
later than it was possible to enter data.

There must be at least one user account in the system:

* either the data was added via the admin, which requires a
  user account
* or it was added via the API, which required an admin account
  to configure the API client credentials

In this edge case, we fall back to the oldest user account as the
owner of the publication/document - it's not expected this will
happen in production environments simply because there aren't
any such environments yet.