:art: [#266] Rework Document API endpoint state transitions
A similar approach like for the publication API endpoints. We call the
state transitions only when there are changes (and in practice, this
means that only the published -> revoked transition can be done
through the API).

Additionally, the publicatiestatus target state validation is actually
model-agnostic, so it's extracted out into a re-usable DRF validator.