:art: [#266] Update existing documents api tests and flow
The majority of document side effects are now triggered via operations
on the related publication, so keeping tests around that change the
publication status is not an accurate reflection of reality.

Additionally, the tests that are still relevant (basically only
the revocation of a document) are moved to the test module that
exclusively deals with these state transitions and their side
effects.