:sparkles: [#215] Calculate GPP app URL from template
The publication model now has a calculated property based on the global
configuration that interpolates the UUID in the provided template.

Note that this may have some performance impact in list pages if the
global configuration needs to be looked up repeatedly - I checked that
the solo cache is enabled and that's the case, through the OAF default
settings. There may still be some overhead for repeatedly accessing the
cache, but if it's a problem, we can optimize this.