:building_construction: [#266] Rework how we process status changes in the admin
The gist is that *before* we call the save method, we determine which
state transitions will need to be called and then we call them *after*
the form save has executed. When creating new objects, it is ensured
that they are first persisted to the database.

There are a number of django-admin behaviours here that work against us,
which we don't suffer from in the serializers:

* the admin doesn't use forms to actually save the data, instead it
  processes them and then wires up everything together. It calls the
  save method with commit=False, and that has a bunch of consequences
  that are... unusual if you're used to plain forms and views
* model forms are *also* weird, apparently as part of their validation
  the existing instance is already being updated with the newly
  submitted field values. we have to *undo* this for the
  publicatiestatus field, otherwise we can't make the state transition
  from source to target because the starting state is not accurate

The code is less elegant than for serializers, and that's for a great
deal because of update/create not being split up.