:memo: [#274] Document the support file-provision flows
Documentation driven development :) The two patterns that we support
and the associated design decisions/new requirements become clear from
the desired mechanisms to provide files and being able to check the
status.

This makes clear that:

* we need a mechanism to check if the 'upload' completed succesfully
  when the download + upload runs in a background job
* we need to better handle the case when someone is too eager to try to
  download a document already before upload is completed

This also creates *some* visibility to the client in case uploads fail,
instead of only showing up in error monitoring.

For now, it's decided that communicating an async task ID and being
able to check the status of that is too complex, but in the future this
may become necessary to give the API clients more control.