:sparkles: [#274] Improve introspection and error handling for incomplete uploads
Attempting to download documents while they're still uploading can be
detected on our end, and we do so now. This will become especially
relevant when the upload is running in a Celery task without easy
way to check the status.

To give the client some indication if the document is ready for upload,
the serializer now includes the property. Clients can opt to poll,
for example.