:recycle: [#274] Refactor existing document serializers
Further split up the document serializer - the base serializer is now
used for the list and detail endpoints (read), and a separate
serializer now exists for the document create. Previously, there was
already a separate serializer for the update operations.

This achieves encapsulation of associated behaviours - e.g. only on
create do we process the file size and sync the document to the
specified Documents API, which results in an array of 'bestandsdelen'
being returned. With update, detail and list operations, this would
always simply return an empty-ish value because only the create
operation has the reference to the remote document in memory.

This *does* modify the API specification and one could argue it's a
breaking change, however behaviour-wise it is not and rather a
clarification of how things were intended to be in the first place.

Note that this also moves the create and update methods to their
respective serializers, improving the locality of related code.

In the end, this makes it easier to add the creation-specific
discriminator field and the future additional field for the document
URL where a document can be retrieved from.