:sparkles: [#274] Update Documenten API client code to retrieve document URL
The client now has utilities to process document downloads in a memory
efficient manner by streaming the download file. The metadata about
expected file upload parts is used to chunk up this download into
in-memory or on-disk (temporary) files that can then be offered for
upload to (another) documents API.

Some code has been refactored since it's now being called in multiple
places and makes sense to be split up a bit more.