:white_check_mark: [#350] Update tests with expected/correct validation behaviour
Relational fields that can be nullable must accept null
as empty value, while not accepting empty strings as they
don't resolve to anything.

Many-to-many and many-to-one relations must accept a list
of references and not empty strings or null.

String-fields must not accept null values for their empty
value, instead a plain string should be provided.