:memo: [#350] Document strict read response schema
The read schema/serializer is effectively returned in to the client
for the following requests:

* publication detail
* (paginated) publication list
* successful create
* successful (partial) update

The help text extensions related to create/update operations
are removed from these components to de-clutter, and the nullable
publisher is now documented and described in which situation it
can be null so that client developers can anticipate this.

Furthermore, all fields returned by the serializer are now marked
as required, which guarantees that a key will be returned and
clients do not need to check for missing keys.

This is all done in the get_fields hook so we get test-time
safety and have less risk of documentation being out of sync
with the actual runtime behaviour, while relying on default
drf-spectacular documentation generation mechanisms.

Next: making the write serializer polymorphic and tuning it
to its situation(s).