:bookmark: [#353] bumped the api version to v2
Bumped the API version to v2 while keep supporting (temporary) v1.
The v1 will be excluded from the documentation, and is filtered out
by the pre_commit_hook remove_excluded_api_paths.

We looked for multiple solutions including the route that the objects
api took. But the versioning route they took looked to complex to
partake for a version which will only be temporarily supported.