:sparkles: [#367] Display an 'add document' link on publication document inline
Solution inspired by Open Zaak template modifications.

It appears that the regular inline behaviour adds the link dynamically
using Javascript. However, we want to display a fixed link to the
document admin instead, which requires modifying the template of the
inline.

Unfortunately, these django templates don't define any blocks so
we must copy them entirely to add an extra table row.

For maintainability reasons (django upgrades), the custom code
is loaded via an include tag to make it clear which part is
custom.